Rename new-trace-viewer to trace-viewer.
Move the directory, rename NewTraceViewer to TraceViewer across web-shared and web, and update the build script and README.
