+++ .changeset/rename-new-trace-viewer.md
@@ -1,0 +1,6 @@
+---
+'@workflow/web-shared': major
+'@workflow/web': major
+---
+
+Rename `new-trace-viewer` to `trace-viewer` and the public `NewTraceViewer` export to `TraceViewer`.
