+++ packages/web-shared/README.md
@@ -13,11 +13,11 @@ If you want a full observability experience with server actions already wired, t
 It comes with pre-styled UI components that accept data + callbacks:
 
 ```tsx
-import { NewTraceViewer } from '@workflow/web-shared';
+import { TraceViewer } from '@workflow/web-shared';
 
 export default function MyRunDetailView({ run, events, fetchSpanDetail }) {
   return (
-    <NewTraceViewer
+    <TraceViewer
       run={run}
       events={events}
       sidebarData={{ run, events, fetchSpanDetail }}
