+++ packages/web-shared/src/components/sidebar/attributes-block.tsx
@@ -8,8 +8,8 @@ import { cva, type VariantProps } from 'class-variance-authority';
 import { ArrowUpRight } from 'lucide-react';
 import type { ReactNode } from 'react';
 import { cn } from '../../lib/cn';
-import { CopyButton } from '../new-trace-viewer/components/copy-button';
-import { MiddleTruncate } from '../new-trace-viewer/components/middle-truncate/middle-truncate';
+import { CopyButton } from '../trace-viewer/components/copy-button';
+import { MiddleTruncate } from '../trace-viewer/components/middle-truncate/middle-truncate';
 import {
   CollapsibleContent,
   CollapsibleRoot,
