+++ packages/web-shared/src/components/trace-viewer.tsx
@@ -2,14 +2,14 @@ import type { WorkflowRun } from '@workflow/core/runtime';
 import type { Event } from '@workflow/world';
 import { useMemo } from 'react';
 import { buildTrace, type TraceWithMeta } from '../lib/trace-builder';
-import { TraceViewerSkeleton } from './new-trace-viewer/components/trace-viewer-skeleton';
-import { NewTraceViewer as NewTraceViewerComponent } from './new-trace-viewer/trace-viewer';
 import {
   type SidebarDataContextValue,
   SidebarDataProvider,
 } from './sidebar/sidebar-data-context';
+import { TraceViewerSkeleton } from './trace-viewer/components/trace-viewer-skeleton';
+import { TraceViewer as TraceViewerComponent } from './trace-viewer/trace-viewer';
 
-const NewTraceViewer = ({
+const TraceViewer = ({
   run,
   events,
   sidebarData,
@@ -41,7 +41,7 @@ const NewTraceViewer = ({
   return (
     <SidebarDataProvider value={sidebarData}>
       <div className="relative w-full h-full flex">
-        <NewTraceViewerComponent
+        <TraceViewerComponent
           trace={trace}
           onLoadMore={onLoadMore}
           hasMore={hasMore}
@@ -52,4 +52,4 @@ const NewTraceViewer = ({
   );
 };
 
-export { NewTraceViewer };
+export { TraceViewer };
