+++ packages/web-shared/src/components/trace-viewer/components/copy-button.tsx
