+++ packages/web-shared/src/components/trace-viewer/components/detail-panel-width.test.ts
