+++ packages/web-shared/src/components/trace-viewer/components/detail-panel.tsx
