+++ packages/web-shared/src/components/trace-viewer/components/draggable-border.tsx
