+++ packages/web-shared/src/components/trace-viewer/components/middle-truncate/copy-selection.ts
