+++ packages/web-shared/src/components/trace-viewer/components/middle-truncate/use-middle-truncate.ts
