+++ packages/web-shared/src/components/trace-viewer/components/minimap.tsx
