+++ packages/web-shared/src/components/trace-viewer/components/pane-constants.ts
