+++ packages/web-shared/src/components/trace-viewer/components/span-markers.tsx
