+++ packages/web-shared/src/components/trace-viewer/components/split-pane.tsx
