+++ packages/web-shared/src/components/trace-viewer/components/trace-shortcut-helper.module.css
