+++ packages/web-shared/src/components/trace-viewer/components/use-element-width.ts
