+++ packages/web-shared/src/components/trace-viewer/components/use-row-window.ts
