+++ packages/web-shared/src/components/trace-viewer/icons.tsx
