+++ packages/web-shared/src/components/trace-viewer/search.ts
