+++ packages/web-shared/src/components/trace-viewer/types.ts
