+++ packages/web-shared/src/components/trace-viewer/util/timing.ts
